refactor: split ocr_reader.py into type/interface/backend/bridge modules
Extract 4 new modules from ocr_reader.py (10,406 -> 7,219 lines):

- ocr_runtime_types.py: all dataclass types, constants, OCR helpers
- ocr_backend_interface.py: CaptureBackend / OcrBackend Protocols
- ocr_rapidocr_backend.py: RapidOcrBackend class
- ocr_bridge_writer.py: OcrReaderBridgeWriter + decorator

Delete dead code ocr_bridge.py (zero imports, 668 lines).
Update test monkeypatches for moved symbols.

938 tests pass, 2 pre-existing failures.
